pyproject.toml
[project]
name = "drawing_mcp"
version = "0.1.0"
description = "MCP drawing utility"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.14.3",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
drawing-mcp = "drawing_mcp.server:main"